requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
bm25s==0.2.7.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cython==3.0.12
distro==1.9.0
faiss-gpu-cu12==1.10.0
filelock==3.17.0
flask==3.1.0
flatbuffers==25.2.10
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
onnxruntime==1.20.1
openai==1.62.0
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic-core==2.27.2
pyjnius==1.6.1
pyserini==0.44.0
pystemmer==2.2.0.3
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
stemmer==0.0.4
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
werkzeug==3.1.3
psutil